requirements.txt
Pillow==10.0.1
PyJWT==1.7.1
Django==5.0.1
django-cors-headers==4.3.1